[struct2tensor][OSS] Pass environment variables to docker compose build script.
This allows the user to pass the environment variables directly into the manylinux build script. The reason we need this is because the version used to build in docker needs to be the same version used to run outside of docker (i.e. the version that presubmits runs).

It also makes docker-compose.yml more maintainable, since we do not need to make a service name for each combination of python/tf version.

PiperOrigin-RevId: 298478995