[struct2tensor] Add visibility flag to s2t_gen_op_wrapper_py.
BEGIN_PUBLIC
Add visibility flag to s2t_gen_op_wrapper_py.
END_PUBLIC
NO_RELEASE_UPDATE=no

The reason I need this is because of the way the parquet directories are structured. The python files are in a separate directory, so the s2t_gen_op_wrapper_py rule needs to be able to specify visibility for the python directory.

The reason this was not an issue before was because previously I was using tf_gen_op_wrapper_py. But I need to switch them all over to s2t's custom rules in order for the code to work in open source.

PiperOrigin-RevId: 300396012